Moving RewardsLibEnv to it's own file
* Also the `create` static method was moved to this class
